+++ splash-screen/src/overlay/image.rs
@@ -32,7 +32,7 @@ impl Overlay for ImageOverlay {
         self.z_index
     }
 
-    fn draw(&self, target: &mut Frame, _timestamp_ms: u64) {
+    fn draw(&self, target: &mut Frame, _timestamp_ms: u128) {
         target.blit(
             self.frame.left as u32,
             self.frame.top as u32,
@@ -42,7 +42,7 @@ impl Overlay for ImageOverlay {
         );
     }
 
-    fn time_to_next_frame_ms(&self, _timestamp_ms: u64) -> Option<u64> {
+    fn time_to_next_frame_ms(&self, _timestamp_ms: u128) -> Option<u128> {
         None
     }
 }
